% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{CalibrationPlots}
\alias{CalibrationPlots}
\title{Generate plots following calibration}
\usage{
CalibrationPlots(wd, countryName)
}
\arguments{
\item{wd}{working directory}

\item{countryName}{string defining country name}
}
\description{
Generate plots following calibration
}

