% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{cluster_summary}
\alias{cluster_summary}
\title{Summary table of clusters}
\usage{
cluster_summary(X, model)
}
\arguments{
\item{X}{Data matrix (obs × vars)}

\item{model}{ClustDeepVar}
}
\value{
Data frame summary
}
\description{
Summary table of clusters
}
