% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{cluster_summary_table}
\alias{cluster_summary_table}
\title{Summary table of cluster composition (Deep or numeric)}
\usage{
cluster_summary_table(X, clusters)
}
\arguments{
\item{X}{Data frame of active variables}

\item{clusters}{Named vector of clusters}
}
\description{
Summary table of cluster composition (Deep or numeric)
}
