% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{compute_centers}
\alias{compute_centers}
\title{Compute cluster centers in latent space (DeepVar)}
\usage{
compute_centers(embeddings, clusters)
}
\arguments{
\item{embeddings}{Matrix (variables × latent_dim)}

\item{clusters}{Vector of cluster indices}
}
\value{
Matrix of cluster centers
}
\description{
Compute cluster centers in latent space (DeepVar)
}
\keyword{internal}
