% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{compute_inertia}
\alias{compute_inertia}
\title{Inertia (within / between / total) for variable embeddings}
\usage{
compute_inertia(embeddings, clusters)
}
\arguments{
\item{embeddings}{Matrix (variables × latent_dim)}

\item{clusters}{Vector of cluster assignments}
}
\value{
List with within, between, total
}
\description{
Inertia (within / between / total) for variable embeddings
}
