% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{compute_silhouette_var}
\alias{compute_silhouette_var}
\title{Silhouette score for variables (correlation distance)}
\usage{
compute_silhouette_var(X, clusters)
}
\arguments{
\item{X}{Original data (observations × variables)}

\item{clusters}{Cluster assignments over variables}
}
\value{
Mean silhouette
}
\description{
Silhouette score for variables (correlation distance)
}
