% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{export_metrics}
\alias{export_metrics}
\title{Export clustering metrics}
\usage{
export_metrics(X, model, file = "deep_metrics")
}
\arguments{
\item{X}{Data matrix}

\item{model}{Deep clustering model}

\item{file}{Base path (without extension)}
}
\description{
Export clustering metrics
}
