% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{importance_deep}
\alias{importance_deep}
\title{Variable importance (DeepVar)}
\usage{
importance_deep(embeddings, clusters)
}
\arguments{
\item{embeddings}{Matrix (variables × latent_dim)}

\item{clusters}{Cluster vector}
}
\value{
Data frame Variable, Cluster, Importance
}
\description{
Importance = 1 / (1 + distance_to_center)
Normalized in \link{0,1}
}
