% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustplot.R
\name{kmeans_hist}
\alias{kmeans_hist}
\title{Title}
\usage{
kmeans_hist(centers, x, y, nstart)
}
\arguments{
\item{centers}{Number of clustering centers}

\item{x}{The dataframe excludes the label vector}

\item{y}{A vector of labels}

\item{nstart}{}
}
\value{
kmeans result
}
\description{
Title
}
\examples{
no example
}
