% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{login}
\alias{login}
\title{Login to DIDE cluster}
\usage{
login(cwd = "~/git/clustr", remotewd = "/tmp/jjo11/clustr",
  cluster = "fi--didemrchnb", log = TRUE)
}
\arguments{
\item{cwd}{set current working directory}

\item{remotewd}{remote directory on network share}

\item{cluster}{select cluster 'fi--didemrchnb' or 'fi--dideclusthn'}

\item{log}{boolean to set logging}
}
\description{
Login to DIDE cluster
}

