% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount.R
\name{mount}
\alias{mount}
\title{Mount network share on local machine (only works for jjo11 currently)}
\usage{
mount(username, remotewd)
}
\arguments{
\item{username}{DIDE username}

\item{remotewd}{remote directory on network share (don't require entire path)}
}
\description{
Mount network share on local machine (only works for jjo11 currently)
}

