% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{open_remote}
\alias{open_remote}
\title{System call to open network share directory}
\usage{
open_remote(username, remotewd)
}
\arguments{
\item{username}{DIDE username}

\item{remotewd}{remote directory on network share (don't require entire path)}
}
\description{
System call to open network share directory
}

