% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{plot_cluster_sizes}
\alias{plot_cluster_sizes}
\title{Plot number of variables per cluster}
\usage{
plot_cluster_sizes(clusters)
}
\arguments{
\item{clusters}{Named vector of cluster assignments}
}
\description{
Plot number of variables per cluster
}
