% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{ProjectionPlots}
\alias{ProjectionPlots}
\title{Generate plots following model projection}
\usage{
ProjectionPlots(wd, countryName)
}
\arguments{
\item{wd}{working directory}

\item{countryName}{string defining country name}
}
\description{
Generate plots following model projection
}

