% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{PushClusterCalibration}
\alias{PushClusterCalibration}
\title{Wrapper around enqueue object to pass calibration function directly to cluster}
\usage{
PushClusterCalibration(countryName, MaxError, MinNumber)
}
\arguments{
\item{countryName}{string defining country name}

\item{MaxError}{maximum acceptable total simulation error}

\item{MinNumber}{minimum number of accepted parameter sets below maxError}
}
\description{
Wrapper around enqueue object to pass calibration function directly to cluster
}

