% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{PushClusterOptimisation}
\alias{PushClusterOptimisation}
\title{Push optimisation directly to cluster [DEVELOPMENT]}
\usage{
PushClusterOptimisation()
}
\description{
Push optimisation directly to cluster [DEVELOPMENT]
}

