% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{RunClusterCalibration}
\alias{RunClusterCalibration}
\title{Cluster calibration function.}
\usage{
RunClusterCalibration(country, data, maxIterations, maxError, limit, parRange,
  targetIterations = 10000)
}
\arguments{
\item{country}{string defining country name}

\item{data}{list of calibration-ready data, generated by GetMasterData()}

\item{maxIterations}{maximum number of model simulations to be generated}

\item{maxError}{maximum acceptable total simulation error}

\item{limit}{minimum number of accepted parameter sets below maxError}

\item{parRange}{vector containing cascade parameter range}

\item{targetIterations}{target number of initial and sensical values that will be generated}
}
\description{
This function controls the entire calibration procedure in CascadeDashboard, and is a port of the
version used in the R-package
}

