% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{submit}
\alias{submit}
\title{Submit job to DIDE cluster}
\usage{
submit(q = obj, x)
}
\arguments{
\item{q}{queue from didewin}

\item{x}{function to run on the cluster}
}
\description{
Submit job to DIDE cluster
}

