% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to DIDE cluster}
\usage{
login(username = "jjo11", remotewd = "clustr", sources = NULL,
  packages = c("cascade", "CascadeDashboard", "devtools"), cluster = "MRC",
  log = TRUE)
}
\arguments{
\item{username}{DIDE username}

\item{remotewd}{remote directory on network share (don't require entire path)}

\item{sources}{vector of required sources}

\item{packages}{vector of required packages}

\item{cluster}{either "MRC" or "DIDE"}

\item{log}{boolean to set logging}
}
\description{
Login to DIDE cluster
}

