% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableClustering.R
\name{VariableClustering}
\alias{VariableClustering}
\title{VariableClustering: Unified Interface for Variable Clustering}
\description{
High-level R6 wrapper providing a scikit-learn style interface for variable
clustering.

This class centralizes access to all clustering algorithms implemented in the
ClustR package.
}
\details{
This class acts as a \emph{controller} that:
\itemize{
\item selects the appropriate clustering algorithm,
\item forwards \code{.fit()} and \code{.predict()} calls,
\item provides unified printing, summarization and plotting methods,
\item ensures API consistency across algorithms.
}
}
\examples{
\dontrun{
library(ClustR)

X <- matrix(rnorm(200 * 12), 200, 12)

model <- VariableClustering$new(
  method = "deep",
  n_clusters = 3,
  latent_dim = 3
)

model$fit(X)
model$summary()
model$plot()

# Access clusters:
model$get_clusters()
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{method}}{Character string: "deep" (others added later)}

\item{\code{model}}{R6 object. The underlying clustering model instance.}

\item{\code{n_clusters}}{Number of clusters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VariableClustering-new}{\code{VariableClustering$new()}}
\item \href{#method-VariableClustering-fit}{\code{VariableClustering$fit()}}
\item \href{#method-VariableClustering-predict}{\code{VariableClustering$predict()}}
\item \href{#method-VariableClustering-get_clusters}{\code{VariableClustering$get_clusters()}}
\item \href{#method-VariableClustering-summary}{\code{VariableClustering$summary()}}
\item \href{#method-VariableClustering-plot}{\code{VariableClustering$plot()}}
\item \href{#method-VariableClustering-to_dataframe}{\code{VariableClustering$to_dataframe()}}
\item \href{#method-VariableClustering-print}{\code{VariableClustering$print()}}
\item \href{#method-VariableClustering-clone}{\code{VariableClustering$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-new"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-new}{}}}
\subsection{Method \code{new()}}{
Constructor for VariableClustering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$new(method = "deep", n_clusters = 3, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Character. Currently only \code{"deep"} is available.}

\item{\code{n_clusters}}{Integer. Number of clusters.}

\item{\code{...}}{Additional arguments forwarded to the selected algorithm.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-fit"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-fit}{}}}
\subsection{Method \code{fit()}}{
Fit clustering model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$fit(X, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Numeric matrix or data frame.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The VariableClustering object (invisible).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-predict"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-predict}{}}}
\subsection{Method \code{predict()}}{
Predict cluster membership for new variables (if supported)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$predict(X_new, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_new}}{New matrix or data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-get_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-get_clusters}{}}}
\subsection{Method \code{get_clusters()}}{
Return vector of cluster assignments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$get_clusters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-summary"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-summary}{}}}
\subsection{Method \code{summary()}}{
Print clustering summary (delegates to underlying model)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-plot"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method depending on selected algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$plot(type = "default", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Export results to a tidy data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$to_dataframe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-print"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-print}{}}}
\subsection{Method \code{print()}}{
Custom print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableClustering-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VariableClustering-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableClustering$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
